Move token generation from AIApp to Config
